requirements.txt
scikit-learn~=1.3.0
pandas~=1.5.2
pyttsx3~=2.90
nltk~=3.8.1
spacy~=3.6.1
python-dotenv~=0.21.0
numpy~=1.23.5